perf(desktop): make message queues instant
